+++ src/JsonBoolean.php
@@ -64,4 +64,9 @@ final class JsonBoolean implements JsonValue
     {
         return $this->boolValue() === $another->boolValue();
     }
+
+    public function isNull(): bool
+    {
+        return false;
+    }
 }